+++ src/styles/css/main.stylesheet.css
@@ -801,10 +801,6 @@ div.icon img {
     width: 200px;
 }
 
-.testbench-runall-label {
-    display: none;
-}
-
 .tb-dialog-button {
     display: inline;
     margin: 8px;
@@ -1856,7 +1852,7 @@ canvas {
     margin-bottom: 0.5rem;
 }
 
-/* 
+/*
 .ProseMirror ul {
     list-style-type: disc;
 }
